Services: Test that simple definitions turn into full definition
There are a lot of abbreviations for defining services,
actions and other WordPress things.
Handling those is a bit hard for the Provider
because there are slight differences
in each of their definition structure.
After some projects we found a better way of defining things
and this is the first step toward it.